+++ glfw/src/key.zig
@@ -214,7 +214,7 @@ pub const Key = enum(c_int) {
     /// @thread_safety This function must only be called from the main thread.
     ///
     /// see also: input_key_name
-    pub inline fn getName(self: Key, scancode: isize) error{PlatformError}!?[:0]const u8 {
+    pub inline fn getName(self: Key, scancode: i32) error{PlatformError}!?[:0]const u8 {
         internal_debug.assertInitialized();
         const name_opt = cc.glfwGetKeyName(@enumToInt(self), @intCast(c_int, scancode));
         getError() catch |err| return switch (err) {
@@ -240,7 +240,7 @@ pub const Key = enum(c_int) {
     /// Possible errors include glfw.Error.NotInitialized, glfw.Error.InvalidEnum and glfw.Error.PlatformError.
     ///
     /// @thread_safety This function may be called from any thread.
-    pub inline fn getScancode(self: Key) error{PlatformError}!isize {
+    pub inline fn getScancode(self: Key) error{PlatformError}!i32 {
         internal_debug.assertInitialized();
         const scancode = cc.glfwGetKeyScancode(@enumToInt(self));
         getError() catch |err| return switch (err) {
